pyproject.toml
[project]
name = "hirehelper"
version = "0.1.0"
description = "A python tool to help answer questions in the hiring forms."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-community>=0.3.25",
    "langchain[google-genai]>=0.3.25",
    "python-dotenv>=1.1.0",
]
